ref(*): move all source code under src/
in preparation to make boot.go multi-purpose
