+++ src/chans/join.go
