+++ src/chans/join_test.go
