+++ src/clustersrv/client_test.go
