+++ src/riak/config/config.go
