+++ src/riak/discovery.go
