+++ src/riak/discovery_test.go
